Wrap service pods in Deployment/Service
This wraps the service pods with deployments to allow for
self healing if a pod crashes. This also adds a NodePort Service
for the jumphost service, and it removes the node port range
check, in favor of relying on the validation of the Service
itself by the kubernetes control plane, and allowing for
auto-allocation if the node port is not set.

Signed-off-by: Sean Eagan <seaneagan1@gmail.com>
Change-Id: I78b0b651765a43f585a790e96c4472b841fc950d