Add Docker image publish job
We rely on Quay.io to publish new images when commits are merged to the
SIP repository. While this has been a less error-prone approach to
publishing our Docker images, it removes control of image publishing
from the hands of SIP developers, as Airship working committee members
are the only ones who can access our Quay repositories. This change
creates a job to publish our images so that the means for doing so is
transparent to developers, reusable for operators downstream, and
introduces tags for repository commits like other Airship repositories
have.

Signed-off-by: Drew Walters <andrew.walters@att.com>
Change-Id: Idb1b405e6f71bd6a99b24b2f0cfb37d9df463ba3