Fix vBMH Secret extrapolation
The scheduling algorithm subtracts from the number of hosts available
for scheduling each time it encounters a problem retrieving BMC auth
credentials or interface IP addresses, sometimes multiple times for the
same host, causing SIP to display an error that not enough hosts are
available for scheduling when there are. This change fixes the logic by
halting processing of a host when SIP encounters a problem. This change
also aggregates the errors so all problem hosts are identified in the
logs.

Fixes #2

Signed-off-by: Drew Walters <andrew.walters@att.com>
Change-Id: Icd45dff7c420abfc91593edf145807cb05f3b472