Fix JumpHost Deployment Order
When SIP applies the JumpHost Deployment to the cluster first, the
containers can get hung up when starting because the ConfigMap and
Secret do not exist yet. Additionally, declaring the JumpHost container
as a separate variable causes some fields to be lost when applying the
Deployment to Kubernetes. This change re-orders the deployment of the
aforementioned objects in order to fix SIP's "greenfield" deployment
scenario.

Signed-off-by: Drew Walters <andrew.walters@att.com>
Change-Id: Icf541e59754f7e530ef8385df8e406883b6b4661