Remove BMH CRD from SIP CRD entrypoint
The BMH CRD should not be bundled with the SIP CRD because it could
cause conflicts with baremetal operator deployments. The reason the CRD
was bundled with the SIP CRD was to load it into the unit tests.

This change moves the BMH CRD near the sample BMH objects for testing
purposes and adjusts the tests to load from that directory.

Signed-off-by: Drew Walters <andrew.walters@att.com>
Change-Id: I7e5b67ebbabb815fd19320d248246055decbde1e